requirements.txt
algoliasearch==2.0.4
toml==0.10.0
colorama==0.4.1
pyyaml==5.1.1